fix: Rework how head categories are handled internally and player heads
The player heads category exists and creates new heads for each player that joins the server.
Probably because of the switch to XSeries's SkullUtils we don't have the skin data in the Login-Event anymore.
Thus the move to creating the head on-join.

There was also an issue with how Category instances are created and stored in a Head instance
**and** in the HeadManager.
And the HeadManager and Head categories were used for filtering and the GUI.

Categories were identified case insensitive in most places while there was no mechanism to prevent duplicates – Now there is a basic one.